refactor(dataReader): refact of dataReader module stub
