+++ src/index.js
@@ -1,25 +1,8 @@
 import './assets/app.scss'
+import readData from './modules/dataReader'
 import renderSidebar from './modules/sidebar'
 
 document.addEventListener('DOMContentLoaded', () => {
-  readData((data) => {
-    renderSidebar(data)
-  })
+  const data = readData()
+  renderSidebar(data)
 })
-
-function readData(callback) {
-  const data = [
-    {
-      status: 200,
-      gender: "Male"
-    },
-    {
-      status: 422,
-      gender: "Female"
-    }
-  ]
-
-  if(callback) {
-    callback(data)
-  }
-}
