+++ src/modules/dataReader/index.js
@@ -1,0 +1,14 @@
+export default () => {
+  const data = [
+    {
+      status: 200,
+      gender: "Male"
+    },
+    {
+      status: 422,
+      gender: "Female"
+    }
+  ]
+
+  return data
+}
